Definition expression handling
Previously in v100.1.0, setting definition expression on a layer only impacted the layer.  With changes in v100.2.0, definition expressions are also applied to the table.  To support this change, the workflow for displaying EMU layers at different depths had to be changed.  The layer is made invisible, the definition expression set to null and then the query is executed.